# openapi.json
{"openapi":"3.1.0","info":{"title":"Ad Service","description":"Test","version":"0.1.0"},"paths":{"/signup":{"post":{"tags":["Auth"],"summary":"Signup","operationId":"signup_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserExtended"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserExtended"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/refresh_tokens":{"post":{"tags":["Auth"],"summary":"Refresh Tokens","operationId":"refresh_tokens_refresh_tokens_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyCookie":[]}]}},"/{user_id}/make-admin":{"post":{"tags":["User"],"summary":"Make Admin","operationId":"make_admin__user_id__make_admin_post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{user_id/ban":{"post":{"tags":["User"],"summary":"Ban User","operationId":"ban_user__user_id_ban_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{user_id/unban":{"post":{"tags":["User"],"summary":"Unban User","operationId":"unban_user__user_id_unban_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/add":{"post":{"tags":["Ad"],"summary":"Add Ad","operationId":"add_ad_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdAdd"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/all":{"get":{"tags":["Ad"],"summary":"Get All Ads","operationId":"get_all_ads_all_get","parameters":[{"name":"desc_sorting","in":"query","required":true,"schema":{"type":"boolean","title":"Desc Sorting"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Per Page"}},{"name":"title","in":"query","required":false,"schema":{"type":"string","title":"Title"}},{"name":"body","in":"query","required":false,"schema":{"type":"string","title":"Body"}},{"name":"type","in":"query","required":false,"schema":{"allOf":[{"$ref":"#/components/schemas/AdType"}],"title":"Type"}},{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdExtended"},"title":"Response Get All Ads All Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{advert_id}":{"get":{"tags":["Ad"],"summary":"Get Ad By Id","operationId":"get_ad_by_id__advert_id__get","parameters":[{"name":"ad_id","in":"query","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdExtended"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Ad"],"summary":"Delete Advert","operationId":"delete_advert__advert_id__delete","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"ad_id","in":"query","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Ad"],"summary":"Update Ad Type","operationId":"update_ad_type__advert_id__patch","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"ad_id","in":"query","required":true,"schema":{"type":"integer","title":"Ad Id"}},{"name":"ad_type","in":"query","required":true,"schema":{"$ref":"#/components/schemas/AdType"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdExtended"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advert/{advert_id}/comment/add":{"post":{"tags":["Comment"],"summary":"Add Comment","operationId":"add_comment_advert__advert_id__comment_add_post","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"ad_id","in":"query","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advert/{advert_id}/comments":{"get":{"tags":["Comment"],"summary":"Get All By Ad Id","operationId":"get_all_by_ad_id_advert__advert_id__comments_get","parameters":[{"name":"ad_id","in":"query","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentExtended"},"title":"Response Get All By Ad Id Advert  Advert Id  Comments Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/comment/{comment_id}":{"delete":{"tags":["Comment"],"summary":"Delete Comment","operationId":"delete_comment_comment__comment_id__delete","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"comment_id","in":"path","required":true,"schema":{"type":"integer","title":"Comment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advert/{advert_id}/complaint/add":{"post":{"tags":["Complaint"],"summary":"Add Complaint","operationId":"add_complaint_advert__advert_id__complaint_add_post","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"ad_id","in":"query","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advert/{advert_id}/complaints":{"get":{"tags":["Complaint"],"summary":"Get All By Ad Id","operationId":"get_all_by_ad_id_advert__advert_id__complaints_get","parameters":[{"name":"ad_id","in":"query","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ComplaintExtended"},"title":"Response Get All By Ad Id Advert  Advert Id  Complaints Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/complaint/{complaint_id}":{"delete":{"tags":["Complaint"],"summary":"Delete","operationId":"delete_complaint__complaint_id__delete","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"complaint_id","in":"path","required":true,"schema":{"type":"integer","title":"Complaint Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AdAdd":{"properties":{"title":{"type":"string","title":"Title"},"body":{"type":"string","title":"Body"},"compensation":{"type":"string","title":"Compensation"},"location":{"type":"string","title":"Location"},"type":{"$ref":"#/components/schemas/AdType"}},"type":"object","required":["title","body","type"],"title":"AdAdd"},"AdExtended":{"properties":{"title":{"type":"string","title":"Title"},"body":{"type":"string","title":"Body"},"compensation":{"type":"string","title":"Compensation"},"location":{"type":"string","title":"Location"},"type":{"$ref":"#/components/schemas/AdType"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"user":{"$ref":"#/components/schemas/UserExtended"}},"type":"object","required":["title","body","type","id","created_at","updated_at","user"],"title":"AdExtended"},"AdType":{"type":"string","enum":["purchase","sell","service"],"title":"AdType"},"CommentCreate":{"properties":{"body":{"type":"string","title":"Body"}},"type":"object","required":["body"],"title":"CommentCreate"},"CommentExtended":{"properties":{"body":{"type":"string","title":"Body"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"created_by":{"type":"integer","title":"Created By"}},"type":"object","required":["body","id","created_at","created_by"],"title":"CommentExtended"},"ComplaintCreate":{"properties":{"body":{"type":"string","title":"Body"}},"type":"object","required":["body"],"title":"ComplaintCreate"},"ComplaintExtended":{"properties":{"body":{"type":"string","title":"Body"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"created_by":{"type":"integer","title":"Created By"},"ad_id":{"type":"integer","title":"Ad Id"}},"type":"object","required":["body","id","created_at","created_by","ad_id"],"title":"ComplaintExtended"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","title":"Email"},"phone":{"type":"string","title":"Phone"}},"type":"object","required":["username","password","full_name","email","phone"],"title":"UserCreate"},"UserExtended":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"created_at":{"type":"string","format":"date","title":"Created At"},"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","title":"Email"},"phone":{"type":"string","title":"Phone"},"is_admin":{"type":"boolean","title":"Is Admin"},"is_banned":{"type":"boolean","title":"Is Banned"}},"type":"object","required":["id","username","created_at","full_name","email","phone","is_admin","is_banned"],"title":"UserExtended"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyCookie":{"type":"apiKey","in":"cookie","name":"access_token"}}}}